Removed ingress-dispatcher, move the remaining ingress-dispatcher logic into ingress-kafka module
