--- crates/ingress-dispatcher/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "restate-ingress-dispatcher"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-publish = false
-
-[features]
-default = []
-test-util = ["dep:restate-test-util"]
-
-[dependencies]
-restate-bifrost = { workspace = true }
-restate-core = { workspace = true }
-# todo: only needed for DedupInformation :-( Probably fixed by merging restate-storage-api with restate-types
-restate-storage-api = { workspace = true }
-restate-test-util = { workspace = true, optional = true }
-restate-types = { workspace = true }
-restate-wal-protocol = { workspace = true }
-
-anyhow = { workspace = true }
-bytes = { workspace = true }
-bytestring = { workspace = true }
-futures = { workspace = true }
-thiserror = { workspace = true }
-tokio = { workspace = true }
-tracing = { workspace = true }
-
-[dev-dependencies]
-restate-bifrost = { workspace = true, features = ["test-util"] }
-restate-core = { workspace = true, features = ["test-util"] }
-restate-test-util = { workspace = true, features = ["prost"] }
-restate-types = { workspace = true, features = ["test-util"] }
-
-googletest = { workspace = true }
-test-log = { workspace = true }
-tracing-subscriber = { workspace = true }
