+++ crates/ingress-kafka/Cargo.toml
@@ -12,11 +12,13 @@ default = []
 options_schema = ["dep:schemars"]
 
 [dependencies]
+restate-bifrost = { workspace = true }
 restate-core = { workspace = true }
 restate-errors = { workspace = true }
-restate-ingress-dispatcher = { workspace = true }
+restate-storage-api = { workspace = true }
 restate-timer-queue = { workspace = true }
 restate-types = { workspace = true }
+restate-wal-protocol = { workspace = true }
 
 anyhow = { workspace = true }
 base64 = { workspace = true }
@@ -33,7 +35,6 @@ tracing = { workspace = true }
 tracing-opentelemetry = { workspace = true }
 
 [dev-dependencies]
-restate-ingress-dispatcher = { workspace = true, features = ["test-util"] }
 restate-test-util = { workspace = true }
 restate-types = { workspace = true, features = ["test-util"] }
