+++ crates/worker/Cargo.toml
@@ -24,7 +24,6 @@ options_schema = [
 restate-bifrost = { workspace = true }
 restate-core = { workspace = true }
 restate-errors = { workspace = true }
-restate-ingress-dispatcher = { workspace = true }
 restate-ingress-http = { workspace = true }
 restate-ingress-kafka = { workspace = true }
 restate-invoker-api = { workspace = true }
